+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to URLs-LE will be documented here.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.1.0] - 2026-08-04
+## [2.1.0] - 2026-08-05
 
 ### Added